marked the class deprecated
although the concept of code fragments that have fitness is interesting,
it made me confused because it is easy to forget where the fitness is
coming from. As a result, I am dropping this class and will favor a
class that does not have this problem. On the other hand, I do not need
this class for my experiments so I think I am better off without it.